prepwise-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
firebase-admin==6.2.0
google-generativeai==0.3.1
PyPDF2==3.0.1
python-docx==1.1.0
opencv-python==4.8.1.78
numpy
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
requests==2.31.0
